src/main/java/com/ym/projects/apis/posapi/services/impl/CompanyServiceImpl.java
package com.ym.projects.apis.posapi.services.impl;

import com.ym.projects.apis.posapi.model.Company;
import com.ym.projects.apis.posapi.repositories.CompanyRepository;
import com.ym.projects.apis.posapi.services.CompanyService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@Slf4j
public class CompanyServiceImpl implements CompanyService {
    private final CompanyRepository companyRepository;

    public CompanyServiceImpl(CompanyRepository companyRepository) {
        this.companyRepository = companyRepository;
    }

    @Override
    public List<Company> findAllCompany() {
        return companyRepository.findAll();
    }

    @Override
    public Company findCompanyById(Long id) {
        return companyRepository.findById(id).get();
    }

    @Override
    public Company saveOrUpdateCompany(Company company) {
        return companyRepository.save(company);
    }

    @Override
    public void deleteCompanyById(Long id) {
        companyRepository.deleteById(id);
    }
}
